move enum to util package
