+++ reservation-service/src/main/java/com/revature/dto/ReservationDTO.java
@@ -1,7 +1,7 @@
 package com.revature.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.revature.model.RoomType;
+import com.revature.util.RoomType;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
