+++ reservation-service/src/main/java/com/revature/model/Reservation.java
@@ -2,7 +2,7 @@ package com.revature.model;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import io.swagger.annotations.ApiModelProperty;
+import com.revature.util.RoomType;
 
 import javax.persistence.*;
 import java.text.ParseException;
