+++ reservation-service/src/main/java/com/revature/util/RoomType.java
@@ -1,4 +1,4 @@
-package com.revature.model;
+package com.revature.util;
 
 public enum RoomType {
