+++ reservation-service/src/test/java/com/revature/service/TestCaliberIntegrations.java
@@ -4,7 +4,7 @@ import static org.junit.Assert.*;
 
 import com.revature.dto.BatchDTO;
 import com.revature.model.Reservation;
-import com.revature.model.RoomType;
+import com.revature.util.RoomType;
 import com.revature.repository.ReservationRepository;
 
 import java.util.Optional;
